fix: moved action to files domain
